Check Feasibility of Alpha Bound for Optimization Problem

This function checks whether a given value of alpha is a feasible solution to a min/max optimization problem using quadratic functions for sigma_x and sigma_y. The function supports different optimization methods and iteratively attempts to solve the problem.

examine_alpha_bound(
  alpha,
  n_person,
  sigma_x_info,
  sigma_y_info,
  alpha_type,
  score_max = 1,
  num_try = 1,
  method = "GA",
  ...
)

Arguments

alpha

A numeric value representing the alpha value to check.

n_person

An integer representing the number of individuals or rows in the data.

sigma_x_info

A list containing the quadratic function information for sigma_x, including:

A matrix representing the quadratic coefficients for sigma_x.

A vector representing the linear coefficients for sigma_x.

A scalar constant term for sigma_x.

sigma_y_info

A list containing the quadratic function information for sigma_y, including:

A matrix representing the quadratic coefficients for sigma_y.

A vector representing the linear coefficients for sigma_y.

A scalar constant term for sigma_y.

alpha_type

A character string indicating whether the problem is to minimize or maximize alpha. It must be either 'min' or 'max'.

score_max

An integer representing the largest possible score for any test item. Default is 1.

num_try

An integer specifying the number of times to run the optimization algorithm. Default is 1.

method

A character string specifying the optimization method to use. Options are 'GA' (Genetic Algorithm), 'DEoptim' (Differential Evolution), and 'nloptr' (Sequential Least Squares Programming). Default is 'GA'.

...

Additional parameters passed to the optimization algorithm.

Value

A list with the following elements:

result

A logical value indicating whether the alpha is feasible (TRUE) or not (FALSE).

x_value

A numeric vector representing the optimal values of the decision variables, or NULL if not feasible.

Details

The function combines quadratic information from sigma_x_info and sigma_y_info to form a new optimization problem. The optimization checks whether the value of alpha is feasible for either a minimization or maximization problem, depending on the value of alpha_type.

The function supports multiple optimization methods, including Genetic Algorithm (GA), Differential Evolution (DEoptim), and Sequential Least Squares Programming (nloptr). Additional control parameters can be passed through the ... argument to fine-tune the optimization process.

For each iteration, the function calls qp_solver with the combined quadratic function and checks whether the objective function's value is feasible (i.e., less than or equal to 0).

See also