Solve Quadratic Programming Problem using DEoptim

This function solves a quadratic programming problem using the Differential Evolution optimization method from the DEoptim package.

qp_solver_DEoptim(
  n,
  A,
  b,
  c,
  x_max = 1,
  print_message = FALSE,
  NP = 100,
  itermax = 100,
  ...
)

Arguments

n

An integer representing the number of decision variables.

A

A matrix representing the quadratic coefficients.

b

A numeric vector representing the linear coefficients.

c

A numeric scalar representing the constant term in the objective function.

x_max

An integer representing the upper bound for the decision variables. Default is 1.

print_message

A logical value indicating whether to print optimization details. Default is FALSE.

NP

An integer specifying the population size for the DEoptim algorithm. Default is 50.

itermax

An integer specifying the maximum number of iterations. Default is 100.

...

Additional control parameters for DEoptim.

Value

A list containing:

f_cd

The optimal objective function value.

x_value

The optimal values of the decision variables.

See also